rename some files for namespacing
