+++ touchpad/kankaku_touchpad.h
