+++ touchpad/kankaku_utils.c
@@ -1,4 +1,4 @@
-#include "utils.h"
+#include "kankaku_utils.h"
 
 #include <Windows.h>
