+++ touchpad/kankaku_utils.h
@@ -1,12 +1,11 @@
 #ifndef __UTILS_H__
 #define __UTILS_H__
 #include <Windows.h>
+#include <tchar.h>
 
 #include <hidusage.h>
 #include <hidpi.h>
 #pragma comment(lib, "hid.lib")
-
-#include <tchar.h>
 
 struct HID_TOUCH_LINK_COL_INFO
 {
